+++ src/components/ids-locale/ids-locale.ts
@@ -531,6 +531,9 @@ class IdsLocale {
     if (!options || options?.dateStyle === 'short' || options?.year === 'numeric') {
       formattedDate = formattedDate.replace(', ', ' ');
     }
+
+    // eslint-disable-next-line no-irregular-whitespace
+    formattedDate = formattedDate.replace(/ /g, ' '); // remove irregular space
     return formattedDate;
   }
 
@@ -599,6 +602,8 @@ class IdsLocale {
     }
 
     range = range.replace('  ', ' ');
+    range = range.replace(' ', ' ');
+    range = range.replace(' ', ' ');
     if (removePeriod) {
       range = range.replace(':00 -', ' -');
     }